Stop fetching completed tasks count from DB

+++ iexec-core-library/src/main/java/com/iexec/core/metric/PlatformMetric.java
@@ -32,7 +32,7 @@ public class PlatformMetric {
     int aliveAvailableCpu;
     int aliveTotalGpu;
     int aliveAvailableGpu;
-    int completedTasks;
+    long completedTasks;
     long dealEventsCount;
     long dealsCount;
     long replayDealsCount;
